Controller restrict access

+++ app/controllers/articles_controller.rb
@@ -1,6 +1,6 @@
 class ArticlesController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
-  before_action :set_article, only: [:show, :edit, :update]
+  before_action :set_article, only: [:show, :edit, :update, :destroy]
 
   def index
     @articles = Article.all
@@ -27,23 +27,36 @@ class ArticlesController < ApplicationController
   end
 
   def edit
+    unless @article.user == current_user
+      flash[:alert] = 'You can only edit your own article.'
+      redirect_to root_path
+    end
   end
 
   def update
-    if @article.update(article_params)
-      flash[:notice] = 'Article has been updated'
-      redirect_to @article
+    unless @article.user == current_user
+      flash[:alert] = 'You can only update your own article.'
+      redirect_to root_path
     else
-      flash.now[:alert] = 'Article has not been updated'
-      render :edit, status: :unprocessable_entity
+      if @article.update(article_params)
+        flash[:notice] = 'Article has been updated'
+        redirect_to @article
+      else
+        flash.now[:alert] = 'Article has not been updated'
+        render :edit, status: :unprocessable_entity
+      end
     end
   end
 
   def destroy
-    @article = Article.find(params[:id])
-    if @article.destroy
-      flash[:notice] = 'Article has been deleted'
+    unless @article.user == current_user
+      flash[:alert] = 'You can only delete your own article.'
       redirect_to root_path
+    else
+      if @article.destroy
+        flash[:notice] = 'Article has been deleted'
+        redirect_to root_path
+      end
     end
   end
